PR comment: move code into own files and fix lowBalance check
